+++ exchangerates/src/main/java/com/roman/kubik/exchangerates/domain/model/ExchangeRates.kt
@@ -1,5 +1,6 @@
 package com.roman.kubik.exchangerates.domain.model
 
 import com.roman.kubik.currency.Currency
+import java.math.BigDecimal
 
-data class ExchangeRates(val baseCurrency: Currency, val rates: Map<Currency, Double>)
+data class ExchangeRates(val baseCurrency: Currency, val rates: Map<Currency, BigDecimal>)
