+++ exchangerates/src/main/java/com/roman/kubik/exchangerates/ui/ExchangeItemCallback.kt
@@ -1,10 +1,11 @@
 package com.roman.kubik.exchangerates.ui
 
 import com.roman.kubik.exchangerates.domain.model.CurrencyRate
+import java.math.BigDecimal
 
 interface ExchangeItemCallback {
 
     fun onResponderChanged(currencyRate: CurrencyRate)
 
-    fun onAmountEdited(currencyRate: CurrencyRate, amount: Double)
+    fun onAmountEdited(currencyRate: CurrencyRate, amount: BigDecimal)
 }
